Ability to add new rows to each section added
Refactored code to put the data of the ShowIconsViewController into and extension instead of in the main class.

Added the ability to add new rows. Not only be tapping on the green plus button but also by allowing tapping on the row while in editing mode.

Functions that were edited

* numberOfRowsInSection
* cellForRowAtIndexPath
* setEditing(_:animated:)
* commitEditingStyle:forRowAtIndexPath
* didSelectRowAtIndexPath