Indexing was added to the tableView
In this commit there was some major rfactoring of the code to remove the ability to move rows and change some things in the Icon.swift class.

This was to test indexing (alphabetically) the icons and allow for a letter scroller on the left hand side.

See last commit for sections not indexed.